+++ visualization/app/codeCharta/ui/ribbonBar/searchPanel/mapTreeView/mapTreeViewNodeSelector/mapTreeViewNode.selector.ts
@@ -3,6 +3,7 @@ import { klona } from "klona"
 
 import { accumulatedDataSelector } from "../../../../../state/selectors/accumulatedData/accumulatedData.selector"
 import { sortingOrderAscendingSelector } from "../../../../../state/store/appSettings/sortingOrderAscending/sortingOrderAscending.selector"
+import { areaMetricSelector } from "../../../../../state/store/dynamicSettings/areaMetric/areaMetric.selector"
 import { sortingOrderSelector } from "../../../../../state/store/dynamicSettings/sortingOption/sortingOrder.selector"
 import { sortNodesInPlace } from "./sortNodesInPlace"
 
@@ -10,8 +11,9 @@ export const mapTreeViewNodeSelector = createSelector(
     accumulatedDataSelector,
     sortingOrderSelector,
     sortingOrderAscendingSelector,
-    (accumulatedData, sortingOrder, sortingOrderAscending) => {
+    areaMetricSelector,
+    (accumulatedData, sortingOrder, sortingOrderAscending, areaMetric) => {
         // use cloned map as it is sorted inline
-        return sortNodesInPlace(klona(accumulatedData.unifiedMapNode), sortingOrder, sortingOrderAscending)
+        return sortNodesInPlace(klona(accumulatedData.unifiedMapNode), sortingOrder, sortingOrderAscending, areaMetric)
     }
 )